Removed dir check for performance reasons
A profiling run of the script (python -m cProfile top250info.py ...)
revealed that checking if each entry of the movie directory is
actually a directory took a lot of time. I think it's safe to assume
that if there is an entry corresponding to a movie name that then it is
actually a directory. Now the script runs much faster.

If later needed, what could be done is do the check but only for the
entries where a name match has already been found.